Move derived sync classes to their own files
