+++ scene_synchronizer.h
@@ -403,202 +403,6 @@ public:
 			const LocalVector<SenderNetAction> &p_actions) {}
 };
 
-class NoNetSynchronizer : public Synchronizer {
-	friend class SceneSynchronizer;
-
-	bool enabled = true;
-	uint32_t frame_count = 0;
-	LocalVector<NetActionProcessor> pending_actions;
-
-public:
-	NoNetSynchronizer(SceneSynchronizer *p_node);
-
-	virtual void clear() override;
-	virtual void process() override;
-	virtual void on_action_triggered(
-			NetUtility::NodeData *p_node_data,
-			NetActionId p_id,
-			const Array &p_arguments,
-			const Vector<int> &p_recipients) override;
-
-	void set_enabled(bool p_enabled);
-	bool is_enabled() const;
-};
-
-class ServerSynchronizer : public Synchronizer {
-	friend class SceneSynchronizer;
-
-	real_t state_notifier_timer = 0.0;
-
-	struct Change {
-		bool not_known_before = false;
-		Set<StringName> uknown_vars;
-		Set<StringName> vars;
-	};
-
-	enum SnapshotGenerationMode {
-		/// The shanpshot will include The NodeId or NodePath and allthe changed variables.
-		SNAPSHOT_GENERATION_MODE_NORMAL,
-		/// The snapshot will include The NodePath only in case it was unknown before.
-		SNAPSHOT_GENERATION_MODE_NODE_PATH_ONLY,
-		/// The snapshot will include The NodePath only.
-		SNAPSHOT_GENERATION_MODE_FORCE_NODE_PATH_ONLY,
-		/// The snapshot will contains everything no matter what.
-		SNAPSHOT_GENERATION_MODE_FORCE_FULL,
-	};
-
-	/// The changes; the order matters because the index is the NetNodeId.
-	LocalVector<Change> changes;
-
-	OAHashMap<int, NetActionSenderInfo> senders_info;
-	OAHashMap<int, uint32_t> peers_next_action_trigger_input_id;
-
-	uint32_t server_actions_count = 0;
-	LocalVector<SenderNetAction> server_actions;
-
-public:
-	ServerSynchronizer(SceneSynchronizer *p_node);
-
-	virtual void clear() override;
-	virtual void process() override;
-	virtual void on_node_added(NetUtility::NodeData *p_node_data) override;
-	virtual void on_node_removed(NetUtility::NodeData *p_node_data) override;
-	virtual void on_variable_added(NetUtility::NodeData *p_node_data, const StringName &p_var_name) override;
-	virtual void on_variable_changed(NetUtility::NodeData *p_node_data, NetVarId p_var_id, const Variant &p_old_value, int p_flag) override;
-	virtual void on_action_triggered(
-			NetUtility::NodeData *p_node_data,
-			NetActionId p_id,
-			const Array &p_arguments,
-			const Vector<int> &p_recipients) override;
-	virtual void on_actions_received(
-			int sender_peer,
-			const LocalVector<SenderNetAction> &p_actions) override;
-
-	void process_snapshot_notificator(real_t p_delta);
-	Vector<Variant> global_nodes_generate_snapshot(bool p_force_full_snapshot) const;
-	void controller_generate_snapshot(const NetUtility::NodeData *p_node_data, bool p_force_full_snapshot, Vector<Variant> &r_snapshot_result) const;
-	void generate_snapshot_node_data(const NetUtility::NodeData *p_node_data, SnapshotGenerationMode p_mode, Vector<Variant> &r_result) const;
-
-	void execute_actions();
-	void send_actions_to_clients();
-
-	void clean_pending_actions();
-	void check_missing_actions();
-};
-
-class ClientSynchronizer : public Synchronizer {
-	friend class SceneSynchronizer;
-
-	NetUtility::NodeData *player_controller_node_data = nullptr;
-	OAHashMap<NetNodeId, NodePath> node_paths;
-
-	NetUtility::Snapshot last_received_snapshot;
-	std::deque<NetUtility::Snapshot> client_snapshots;
-	std::deque<NetUtility::Snapshot> server_snapshots;
-	uint32_t last_checked_input = 0;
-	bool enabled = true;
-	bool want_to_enable = false;
-
-	bool need_full_snapshot_notified = false;
-
-	struct EndSyncEvent {
-		NetUtility::NodeData *node_data;
-		NetVarId var_id;
-		Variant old_value;
-
-		bool operator<(const EndSyncEvent &p_other) const {
-			if (node_data->id == p_other.node_data->id) {
-				return var_id < p_other.var_id;
-			} else {
-				return node_data->id < p_other.node_data->id;
-			}
-		}
-	};
-
-	Set<EndSyncEvent> sync_end_events;
-
-	uint32_t locally_triggered_actions_count = 0;
-	uint32_t actions_input_id = 0;
-	LocalVector<SenderNetAction> pending_actions;
-	NetActionSenderInfo server_sender_info;
-
-public:
-	ClientSynchronizer(SceneSynchronizer *p_node);
-
-	virtual void clear() override;
-
-	virtual void process() override;
-	virtual void on_node_added(NetUtility::NodeData *p_node_data) override;
-	virtual void on_node_removed(NetUtility::NodeData *p_node_data) override;
-	virtual void on_variable_changed(NetUtility::NodeData *p_node_data, NetVarId p_var_id, const Variant &p_old_value, int p_flag) override;
-	virtual void on_controller_reset(NetUtility::NodeData *p_node_data) override;
-	virtual void on_action_triggered(
-			NetUtility::NodeData *p_node_data,
-			NetActionId p_id,
-			const Array &p_arguments,
-			const Vector<int> &p_recipients) override;
-	virtual void on_actions_received(
-			int sender_peer,
-			const LocalVector<SenderNetAction> &p_actions) override;
-
-	void receive_snapshot(Variant p_snapshot);
-	bool parse_sync_data(
-			Variant p_snapshot,
-			void *p_user_pointer,
-			void (*p_node_parse)(void *p_user_pointer, NetUtility::NodeData *p_node_data),
-			void (*p_input_id_parse)(void *p_user_pointer, uint32_t p_input_id),
-			void (*p_controller_parse)(void *p_user_pointer, NetUtility::NodeData *p_node_data),
-			void (*p_variable_parse)(void *p_user_pointer, NetUtility::NodeData *p_node_data, NetVarId p_var_id, const Variant &p_value));
-
-	void set_enabled(bool p_enabled);
-
-private:
-	/// Store node data organized per controller.
-	void store_snapshot();
-
-	void store_controllers_snapshot(
-			const NetUtility::Snapshot &p_snapshot,
-			std::deque<NetUtility::Snapshot> &r_snapshot_storage);
-
-	void process_controllers_recovery(real_t p_delta);
-
-	void __pcr__fetch_recovery_info(
-			const uint32_t p_input_id,
-			bool &r_need_recover,
-			bool &r_recover_controller,
-			LocalVector<NetUtility::NodeData *> &r_nodes_to_recover,
-			LocalVector<NetUtility::PostponedRecover> &r_postponed_recover);
-
-	void __pcr__sync_pre_rewind(
-			const LocalVector<NetUtility::NodeData *> &p_nodes_to_recover);
-
-	void __pcr__rewind(
-			real_t p_delta,
-			const uint32_t p_checkable_input_id,
-			NetworkedController *p_controller,
-			PlayerController *p_player_controller,
-			const bool p_recover_controller,
-			const LocalVector<NetUtility::NodeData *> &p_nodes_to_recover);
-
-	void __pcr__sync_no_rewind(
-			const LocalVector<NetUtility::PostponedRecover> &p_postponed_recover);
-
-	void apply_last_received_server_snapshot();
-	void process_paused_controller_recovery(real_t p_delta);
-	bool parse_snapshot(Variant p_snapshot);
-	bool compare_vars(
-			const NetUtility::NodeData *p_synchronizer_node_data,
-			const Vector<NetUtility::Var> &p_server_vars,
-			const Vector<NetUtility::Var> &p_client_vars,
-			Vector<NetUtility::Var> &r_postponed_recover);
-
-	void notify_server_full_snapshot_is_needed();
-
-	void send_actions_to_server();
-	void clean_pending_actions();
-	void check_missing_actions();
-};
-
 VARIANT_ENUM_CAST(NetEventFlag)
 
 #endif
